dstribute tests and test deepMerge alternative
